requirements.txt
streamlit
openai
gitpython
python-dotenv
pathlib
chardet